requirements-dev.txt
@# Development dependencies for Face Recognition Attendance System

# Include production requirements
-r requirements.txt

# Development tools
django-extensions==3.2.3
ipython==8.14.0
pytest==7.4.2
pytest-django==4.5.2
black==23.7.0
flake8==6.0.0
